requirements/base.txt
beautifulsoup4==4.3.2
celery==3.1.16
Django==1.7.1
django-celery==3.1.16
django-extensions==1.4.6
django-json-field==0.5.7
django-uuidfield==0.5.0
IPython==2.3.0
Jinja2==2.7.3
psycopg2==2.5.4
python-memcached==1.53
pyzmq==14.4.1
redis==2.10.3
requests==2.4.3
six==1.8.0
tornado==4.0.2
Werkzeug==0.9.6
djangorestframework==2.4.4
